+++ Blender/package.json
@@ -1,0 +1,17 @@
+{
+  "name": "Blender_JSON_Exporter",
+  "version": "6.0.0",
+  "description": "Writes .babylon file from Blender file",
+  "main": "",
+  "repository": {
+    "url": "https://github.com/BabylonJS/Exporters"
+  },
+  "readme": "https://github.com/BabylonJS/Exporters/blob/master/Blender/readme.md",
+  "license": "(Apache-2.0)",
+  "scripts": {
+    "Blender-ZIP": "node buildZipFile.js"
+  },
+  "devDependencies": {
+    "zip-folder": "^1.0.0"
+  }
+}
