+++ Blender/src/babylon_js/logging.py
@@ -13,6 +13,7 @@ class Logger:
     def __init__(self, filename):
         self.start_time = time()
         self.nWarnings = 0
+        self.nErrors   = 0
 
         self.log_handler = open(filename, 'w', encoding='utf8')
         self.log_handler.write('Exporter version: ' + format_exporter_version() + ', Blender version: ' + app.version_string + '\n')
@@ -38,6 +39,11 @@ class Logger:
 
         self.log_handler.close()
         Logger.instance = None
+
+    @staticmethod
+    def error(msg):
+        Logger.log('\nERROR: ' + msg.upper() + '\n', 0, False)
+        Logger.instance.nErrors += 1
 
     @staticmethod
     def warn(msg, numTabIndent = 1, noNewLine = False):
