+++ Blender/src/babylon_js/materials/nodes/fresnel.py
@@ -1,0 +1,10 @@
+from .abstract import AbstractBJSNode
+
+#===============================================================================
+class FresnelBJSNode(AbstractBJSNode):
+    bpyType = 'ShaderNodeFresnel'
+
+    def __init__(self, bpyNode, socketName):
+        super().__init__(bpyNode, socketName)
+
+        self.indexOfRefraction = self.findInput('IOR')
