+++ Blender/src/babylon_js/materials/nodes/mapping.py
@@ -1,0 +1,12 @@
+from .abstract import AbstractBJSNode
+
+#===============================================================================
+class MappingBJSNode(AbstractBJSNode):
+    bpyType = 'ShaderNodeMapping'
+
+    def __init__(self, bpyNode, socketName):
+        super().__init__(bpyNode, socketName)
+
+        self.offset = bpyNode.translation
+        self.ang    = bpyNode.rotation
+        self.scale  = bpyNode.scale
