+++ Blender/src/babylon_js/sound.py
@@ -1,6 +1,5 @@
 from .package_level import *
 
-import bpy
 #===============================================================================
 class Sound:
     def __init__(self, name, autoplay, loop, connectedMesh = None):
@@ -11,7 +10,7 @@ class Sound:
             self.connectedMeshId = connectedMesh.name
             self.maxDistance = connectedMesh.data.maxSoundDistance
 # - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
-    def to_scene_file(self, file_handler):
+    def to_json_file(self, file_handler):
         file_handler.write('{')
         write_string(file_handler, 'name', self.name, True)
         write_bool(file_handler, 'autoplay', self.autoplay)
